backend/pyproject.toml
[project]
name = "mirofish-trading"
version = "0.1.0"
description = "MiroFish-adapted ES futures trading simulation with LLM agents"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "openai>=1.0.0",
    "zep-cloud>=3.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "databento>=0.43.0",
    "google-cloud-storage>=2.14.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
mirofish-sim = "scripts.run_simulation:main"
